ghost/sop/py/requirements.txt
asyncpg==0.27.0
boto3==1.26.23
fastapi==0.88.0
oauthlib==3.2.2
loguru==0.6.0
pandas==1.5.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
python-multipart==0.0.5
requests==2.28.1
SQLAlchemy==1.4.44
uvicorn==0.20.0
uvloop==0.17.0
Authlib==1.2.0
itsdangerous==2.1.2
PyJWT==2.6.0
httpx==0.23.3
loguru-notification